+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+Breaking
+  * AZ-644: Drop Terraform 0.12 support
+
+Added:
+  * AZ-644: Output FrontDoor IP ranges
+
 # v4.4.0 - 2021-12-03
 
 Breaking